feat: boks med animasjon mens BKV3 laster data
